+++ src/senaite/core/upgrade/v02_06_000.py
@@ -2706,3 +2706,16 @@ def migrate_worksheettemplates_to_dx(tool):
             logger.warn("Cannot remove {}. Is not empty".format(origin))
 
     logger.info("Convert Worksheet Templates to Dexterity [DONE]")
+
+
+def reindex_specs(tool):
+    """Reindex the sampletype_uid and sampletype_title indexes from setup
+    catalog for AnalysisSpecs types
+    """
+    logger.info("Reindexing analysis specifications ...")
+    cat = api.get_tool(SETUP_CATALOG)
+    for brain in cat(portal_type="AnalysisSpec"):
+        obj = brain.getObject()
+        logger.info("Reindex analysis spec: %r" % obj)
+        obj.reindexObject(idxs=["sampletype_uid", "sampletype_title"])
+    logger.info("Reindexing analysis specifications [DONE]")